Move commands interfaces to a new package
