+++ src/main/java/redis/clients/jedis/BinaryJedisCluster.java
@@ -1,5 +1,8 @@
 package redis.clients.jedis;
 
+import redis.clients.jedis.commands.BinaryJedisClusterCommands;
+import redis.clients.jedis.commands.JedisClusterBinaryScriptingCommands;
+import redis.clients.jedis.commands.MultiKeyBinaryJedisClusterCommands;
 import redis.clients.jedis.params.set.SetParams;
 import redis.clients.util.KeyMergeUtil;
 import redis.clients.util.SafeEncoder;
@@ -13,7 +16,7 @@ import java.util.Set;
 import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
 
 public class BinaryJedisCluster implements BinaryJedisClusterCommands,
-    MultiKeyBinaryJedisClusterCommands, JedisClusterBinaryScriptingCommands, Closeable {
+        MultiKeyBinaryJedisClusterCommands, JedisClusterBinaryScriptingCommands, Closeable {
 
   public static final short HASHSLOTS = 16384;
   protected static final int DEFAULT_TIMEOUT = 2000;
