+++ src/main/java/redis/clients/jedis/BinaryShardedJedis.java
@@ -7,13 +7,14 @@ import java.util.Set;
 import java.util.regex.Pattern;
 
 import redis.clients.jedis.BinaryClient.LIST_POSITION;
+import redis.clients.jedis.commands.BinaryJedisCommands;
 import redis.clients.jedis.exceptions.JedisConnectionException;
 import redis.clients.jedis.params.set.SetParams;
 import redis.clients.util.Hashing;
 import redis.clients.util.Sharded;
 
 public class BinaryShardedJedis extends Sharded<Jedis, JedisShardInfo> implements
-    BinaryJedisCommands {
+        BinaryJedisCommands {
   public BinaryShardedJedis(List<JedisShardInfo> shards) {
     super(shards);
   }
