+++ src/main/java/redis/clients/jedis/MultiKeyPipelineBase.java
@@ -1,12 +1,14 @@
 package redis.clients.jedis;
 
+import redis.clients.jedis.commands.*;
+
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 public abstract class MultiKeyPipelineBase extends PipelineBase implements
-    MultiKeyBinaryRedisPipeline, MultiKeyCommandsPipeline, ClusterPipeline,
-    BinaryScriptingCommandsPipeline, ScriptingCommandsPipeline {
+        MultiKeyBinaryRedisPipeline, MultiKeyCommandsPipeline, ClusterPipeline,
+        BinaryScriptingCommandsPipeline, ScriptingCommandsPipeline {
 
   protected Client client = null;
