+++ src/main/java/redis/clients/jedis/commands/AdvancedJedisCommands.java
@@ -1,4 +1,4 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
 
 import java.util.List;
