+++ src/main/java/redis/clients/jedis/commands/BinaryJedisClusterCommands.java
@@ -1,16 +1,16 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
+
+import redis.clients.jedis.BinaryClient;
+import redis.clients.jedis.SortingParams;
+import redis.clients.jedis.Tuple;
+import redis.clients.jedis.params.set.SetParams;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import redis.clients.jedis.params.set.SetParams;
-
-/**
- * Common interface for sharded and non-sharded BinaryJedis
- */
-public interface BinaryJedisCommands {
+public interface BinaryJedisClusterCommands {
   String set(byte[] key, byte[] value);
 
   String set(byte[] key, byte[] value, SetParams params);
@@ -213,7 +213,7 @@ public interface BinaryJedisCommands {
 
   Long zremrangeByLex(final byte[] key, final byte[] min, final byte[] max);
 
-  Long linsert(byte[] key, Client.LIST_POSITION where, byte[] pivot, byte[] value);
+  Long linsert(byte[] key, BinaryClient.LIST_POSITION where, byte[] pivot, byte[] value);
 
   Long lpushx(byte[] key, byte[]... arg);
 
@@ -222,8 +222,6 @@ public interface BinaryJedisCommands {
   Long del(byte[] key);
 
   byte[] echo(byte[] arg);
-
-  Long move(byte[] key, int dbIndex);
 
   Long bitcount(final byte[] key);
