+++ src/main/java/redis/clients/jedis/commands/Commands.java
@@ -1,8 +1,12 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
 
 import java.util.Map;
 
 import redis.clients.jedis.BinaryClient.LIST_POSITION;
+import redis.clients.jedis.BitOP;
+import redis.clients.jedis.ScanParams;
+import redis.clients.jedis.SortingParams;
+import redis.clients.jedis.ZParams;
 import redis.clients.jedis.params.set.SetParams;
 
 public interface Commands {
