+++ src/main/java/redis/clients/jedis/commands/JedisClusterBinaryScriptingCommands.java
@@ -1,29 +1,27 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
 
 import java.util.List;
 
-public interface BinaryScriptingCommands {
-
+public interface JedisClusterBinaryScriptingCommands {
   Object eval(byte[] script, byte[] keyCount, byte[]... params);
 
   Object eval(byte[] script, int keyCount, byte[]... params);
 
   Object eval(byte[] script, List<byte[]> keys, List<byte[]> args);
 
-  Object eval(byte[] script);
+  Object eval(byte[] script, byte[] key);
 
-  Object evalsha(byte[] script);
+  Object evalsha(byte[] script, byte[] key);
 
   Object evalsha(byte[] sha1, List<byte[]> keys, List<byte[]> args);
 
   Object evalsha(byte[] sha1, int keyCount, byte[]... params);
 
-  // TODO: should be Boolean, add singular version
-  List<Long> scriptExists(byte[]... sha1);
+  List<Long> scriptExists(byte[] key, byte[][] sha1);
 
-  byte[] scriptLoad(byte[] script);
+  byte[] scriptLoad(byte[] script, byte[] key);
 
-  String scriptFlush();
+  String scriptFlush(byte[] key);
 
-  String scriptKill();
+  String scriptKill(byte[] key);
 }
