+++ src/main/java/redis/clients/jedis/commands/MultiKeyBinaryCommands.java
@@ -1,14 +1,25 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
+
+import redis.clients.jedis.BinaryJedisPubSub;
+import redis.clients.jedis.BitOP;
+import redis.clients.jedis.SortingParams;
+import redis.clients.jedis.ZParams;
 
 import java.util.List;
 import java.util.Set;
 
-public interface MultiKeyBinaryJedisClusterCommands {
+public interface MultiKeyBinaryCommands {
   Long del(byte[]... keys);
 
   List<byte[]> blpop(int timeout, byte[]... keys);
 
   List<byte[]> brpop(int timeout, byte[]... keys);
+
+  List<byte[]> blpop(byte[]... args);
+
+  List<byte[]> brpop(byte[]... args);
+
+  Set<byte[]> keys(byte[] pattern);
 
   List<byte[]> mget(byte[]... keys);
 
@@ -40,6 +51,10 @@ public interface MultiKeyBinaryJedisClusterCommands {
 
   Long sunionstore(byte[] dstkey, byte[]... keys);
 
+  String watch(byte[]... keys);
+
+  String unwatch();
+
   Long zinterstore(byte[] dstkey, byte[]... sets);
 
   Long zinterstore(byte[] dstkey, ZParams params, byte[]... sets);
@@ -55,6 +70,8 @@ public interface MultiKeyBinaryJedisClusterCommands {
   void subscribe(BinaryJedisPubSub jedisPubSub, byte[]... channels);
 
   void psubscribe(BinaryJedisPubSub jedisPubSub, byte[]... patterns);
+
+  byte[] randomBinaryKey();
 
   Long bitop(BitOP op, final byte[] destKey, byte[]... srcKeys);
