+++ src/main/java/redis/clients/jedis/commands/MultiKeyBinaryRedisPipeline.java
@@ -1,4 +1,9 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
+
+import redis.clients.jedis.BitOP;
+import redis.clients.jedis.Response;
+import redis.clients.jedis.SortingParams;
+import redis.clients.jedis.ZParams;
 
 import java.util.List;
 import java.util.Set;
