+++ src/main/java/redis/clients/jedis/commands/MultiKeyCommands.java
@@ -1,4 +1,6 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
+
+import redis.clients.jedis.*;
 
 import java.util.List;
 import java.util.Set;
