+++ src/main/java/redis/clients/jedis/commands/ProtocolCommand.java
@@ -1,4 +1,4 @@
-package redis.clients.jedis;
+package redis.clients.jedis.commands;
 
 public interface ProtocolCommand {
